move subplant identificaiton code
